+++ src/DataCollector/CacheCollector.php
@@ -4,11 +4,20 @@ namespace Barryvdh\Debugbar\DataCollector;
 
 use DebugBar\DataCollector\TimeDataCollector;
 use DebugBar\DataFormatter\HasDataFormatter;
-use Illuminate\Cache\Events\CacheEvent;
-use Illuminate\Cache\Events\CacheHit;
-use Illuminate\Cache\Events\CacheMissed;
-use Illuminate\Cache\Events\KeyForgotten;
-use Illuminate\Cache\Events\KeyWritten;
+use Illuminate\Cache\Events\{
+    CacheFlushed,
+    CacheFlushFailed,
+    CacheFlushing,
+    CacheHit,
+    CacheMissed,
+    ForgettingKey,
+    KeyForgetFailed,
+    KeyForgotten,
+    KeyWriteFailed,
+    KeyWritten,
+    RetrievingKey,
+    WritingKey,
+};
 use Illuminate\Events\Dispatcher;
 
 class CacheCollector extends TimeDataCollector
@@ -19,26 +28,32 @@ class CacheCollector extends TimeDataCollector
     protected $collectValues;
 
     /** @var array */
+    protected $eventStarts = [];
+
+    /** @var array */
     protected $classMap = [
-        CacheHit::class => 'hit',
-        CacheMissed::class => 'missed',
-        KeyWritten::class => 'written',
-        KeyForgotten::class => 'forgotten',
+        CacheHit::class => ['hit', RetrievingKey::class],
+        CacheMissed::class => ['missed', RetrievingKey::class],
+        CacheFlushed::class => ['flushed', CacheFlushing::class],
+        CacheFlushFailed::class => ['flush_failed', CacheFlushing::class],
+        KeyWritten::class => ['written', WritingKey::class],
+        KeyWriteFailed::class => ['write_failed', WritingKey::class],
+        KeyForgotten::class => ['forgotten', ForgettingKey::class],
+        KeyForgetFailed::class => ['forget_failed', ForgettingKey::class],
     ];
 
     public function __construct($requestStartTime, $collectValues)
     {
-        parent::__construct();
+        parent::__construct($requestStartTime);
 
         $this->collectValues = $collectValues;
     }
 
-    public function onCacheEvent(CacheEvent $event)
+    public function onCacheEvent($event)
     {
         $class = get_class($event);
         $params = get_object_vars($event);
-
-        $label = $this->classMap[$class];
+        $label = $this->classMap[$class][0];
 
         if (isset($params['value'])) {
             if ($this->collectValues) {
@@ -52,8 +67,7 @@ class CacheCollector extends TimeDataCollector
             }
         }
 
-
-        if (!empty($params['key']) && in_array($label, ['hit', 'written'])) {
+        if (!empty($params['key'] ?? null) && in_array($label, ['hit', 'written'])) {
             $params['delete'] = route('debugbar.cache.delete', [
                 'key' => urlencode($params['key']),
                 'tags' => !empty($params['tags']) ? json_encode($params['tags']) : '',
@@ -61,14 +75,37 @@ class CacheCollector extends TimeDataCollector
         }
 
         $time = microtime(true);
-        $this->addMeasure($label . "\t" . $event->key, $time, $time, $params);
+        $startHashKey = $this->getEventHash($this->classMap[$class][1] ?? '', $params);
+        $startTime = $this->eventStarts[$startHashKey] ?? $time;
+        $this->addMeasure($label . "\t" . ($params['key'] ?? ''), $startTime, $time, $params);
     }
 
+    public function onStartCacheEvent($event)
+    {
+        $startHashKey = $this->getEventHash(get_class($event), get_object_vars($event));
+        $this->eventStarts[$startHashKey] = microtime(true);
+    }
+
+    private function getEventHash(string $class, array $params): string
+    {
+        unset($params['value']);
+
+        return $class . ':' . substr(hash('sha256', json_encode($params)), 0, 12);
+    }
 
     public function subscribe(Dispatcher $dispatcher)
     {
-        foreach ($this->classMap as $eventClass => $type) {
+        foreach (array_keys($this->classMap) as $eventClass) {
             $dispatcher->listen($eventClass, [$this, 'onCacheEvent']);
+        }
+
+        $startEvents = array_unique(array_filter(array_map(
+            fn ($values) => $values[1] ?? null,
+            array_values($this->classMap)
+        )));
+
+        foreach ($startEvents as $eventClass) {
+            $dispatcher->listen($eventClass, [$this, 'onStartCacheEvent']);
         }
     }