+++ src/LaravelDebugbar.php
@@ -125,8 +125,9 @@ class LaravelDebugbar extends DebugBar
         $this->is_lumen = Str::contains($this->version, 'Lumen');
         if ($this->is_lumen) {
             $this->version = Str::betweenFirst($app->version(), '(', ')');
+        } else {
+            $this->setRequestIdGenerator(new RequestIdGenerator());
         }
-        $this->setRequestIdGenerator(new RequestIdGenerator());
     }
 
     /**
@@ -209,13 +210,16 @@ class LaravelDebugbar extends DebugBar
 
         if ($this->shouldCollect('time', true)) {
             $startTime = $app['request']->server('REQUEST_TIME_FLOAT');
-            $this->addCollector(new TimeDataCollector($startTime));
+
+            if (!$this->hasCollector('time')) {
+                $this->addCollector(new TimeDataCollector($startTime));
+            }
 
             if ($config->get('debugbar.options.time.memory_usage')) {
                 $this['time']->showMemoryUsage();
             }
 
-            if ($startTime) {
+            if ($startTime && !$this->isLumen()) {
                 $app->booted(
                     function () use ($startTime) {
                         $this->addMeasure('Booting', $startTime, microtime(true), [], 'time');
@@ -276,7 +280,8 @@ class LaravelDebugbar extends DebugBar
             try {
                 $startTime = $app['request']->server('REQUEST_TIME_FLOAT');
                 $collectData = $config->get('debugbar.options.events.data', false);
-                $this->addCollector(new EventCollector($startTime, $collectData));
+                $excludedEvents = $config->get('debugbar.options.events.excluded', []);
+                $this->addCollector(new EventCollector($startTime, $collectData, $excludedEvents));
                 $events->subscribe($this['event']);
             } catch (Exception $e) {
                 $this->addCollectorException('Cannot add EventCollector', $e);
@@ -358,6 +363,11 @@ class LaravelDebugbar extends DebugBar
             $queryCollector->setLimits($config->get('debugbar.options.db.soft_limit'), $config->get('debugbar.options.db.hard_limit'));
             $queryCollector->setDurationBackground($config->get('debugbar.options.db.duration_background'));
 
+            $threshold = $config->get('debugbar.options.db.slow_threshold', false);
+            if ($threshold && !$config->get('debugbar.options.db.only_slow_queries', true)) {
+                $queryCollector->setSlowThreshold($threshold);
+            }
+
             if ($config->get('debugbar.options.db.with_params')) {
                 $queryCollector->setRenderSqlWithParams(true);
             }
@@ -397,9 +407,10 @@ class LaravelDebugbar extends DebugBar
                             return; // Issue 776 : We've turned off collecting after the listener was attached
                         }
 
-                        //allow collecting only queries slower than a specified amount of milliseconds
                         $threshold = app('config')->get('debugbar.options.db.slow_threshold', false);
-                        if (!$threshold || $query->time > $threshold) {
+                        $onlyThreshold = app('config')->get('debugbar.options.db.only_slow_queries', true);
+                        //allow collecting only queries slower than a specified amount of milliseconds
+                        if (!$onlyThreshold || !$threshold || $query->time > $threshold) {
                             $this['queries']->addQuery($query);
                         }
                     }
@@ -470,11 +481,16 @@ class LaravelDebugbar extends DebugBar
         if ($this->shouldCollect('models', true) && $events) {
             try {
                 $this->addCollector(new ObjectCountCollector('models'));
-                $events->listen('eloquent.retrieved:*', function ($event, $models) {
-                    foreach (array_filter($models) as $model) {
-                        $this['models']->countClass($model);
-                    }
-                });
+                $eventList = ['retrieved', 'created', 'updated', 'deleted'];
+                $this['models']->setKeyMap(array_combine($eventList, array_map('ucfirst', $eventList)));
+                $this['models']->collectCountSummary(true);
+                foreach ($eventList as $event) {
+                    $events->listen("eloquent.{$event}: *", function ($event, $models) {
+                        $event = explode(': ', $event);
+                        $count = count(array_filter($models));
+                        $this['models']->countClass($event[1], $count, explode('.', $event[0])[1]);
+                    });
+                }
             } catch (Exception $e) {
                 $this->addCollectorException('Cannot add Models Collector', $e);
             }
@@ -566,6 +582,11 @@ class LaravelDebugbar extends DebugBar
         if ($this->shouldCollect('gate', false)) {
             try {
                 $this->addCollector($app->make(GateCollector::class));
+
+                if ($config->get('debugbar.options.gate.trace', false)) {
+                    $this['gate']->collectFileTrace(true);
+                    $this['gate']->addBacktraceExcludePaths($config->get('debugbar.options.gate.exclude_paths',[]));
+                }
             } catch (Exception $e) {
                 $this->addCollectorException('Cannot add GateCollector', $e);
             }
@@ -656,7 +677,10 @@ class LaravelDebugbar extends DebugBar
      */
     public function handleError($level, $message, $file = '', $line = 0, $context = [])
     {
-        $this->addThrowable(new \ErrorException($message, 0, $level, $file, $line));
+        if ($this->hasCollector('exceptions')) {
+            $this['exceptions']->addWarning($level, $message, $file, $line);
+        }
+
         if ($this->hasCollector('messages')) {
             $file = $file ? ' on ' . $this['messages']->normalizeFilePath($file) . ":{$line}" : '';
             $this['messages']->addMessage($message . $file, 'deprecation');
@@ -740,7 +764,7 @@ class LaravelDebugbar extends DebugBar
         $this->addThrowable(
             new Exception(
                 $message . ' on Laravel Debugbar: ' . $exception->getMessage(),
-                $exception->getCode(),
+                (int) $exception->getCode(),
                 $exception
             )
         );