+++ CHANGELOG.md
@@ -1,5 +1,66 @@
 # Changelog
 
+## v3.16.2 - 2025-12-16
+
+### What's Changed
+
+* Remove default null value env by @Erulezz in https://github.com/barryvdh/laravel-debugbar/pull/1815
+* Remove --ignore-platform-req=php+ on integration test setup by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1814
+* Remove calls to PHP 8.5-deprecated `setAccessible` by @jnoordsij in https://github.com/barryvdh/laravel-debugbar/pull/1822
+
+### New Contributors
+
+* @Erulezz made their first contribution in https://github.com/barryvdh/laravel-debugbar/pull/1815
+
+**Full Changelog**: https://github.com/barryvdh/laravel-debugbar/compare/v3.16.1...v3.16.2
+
+## v3.16.1 - 2025-11-19
+
+### What's Changed
+
+* Slow threshold highlight on queries by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1805
+* (fix) trim last line breaks on logs by @angeljqv in https://github.com/barryvdh/laravel-debugbar/pull/1806
+* fix: Typo by @aurac in https://github.com/barryvdh/laravel-debugbar/pull/1810
+* Test on PHP 8.5 by @jnoordsij in https://github.com/barryvdh/laravel-debugbar/pull/1811
+* Add '_boost*' to debugbar exceptions by @barryvdh in https://github.com/barryvdh/laravel-debugbar/pull/1818
+* Dropped Laravel 9 support
+
+### New Contributors
+
+* @aurac made their first contribution in https://github.com/barryvdh/laravel-debugbar/pull/1810
+
+**Full Changelog**: https://github.com/barryvdh/laravel-debugbar/compare/v3.16.0...v3.16.1
+
+## v3.16.0 - 2025-07-21
+
+### What's Changed
+
+* Make all scalar config values configurable through environment variables by @wimski in https://github.com/barryvdh/laravel-debugbar/pull/1784
+* Check if file exists on FilesystemStorage by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1790
+* Bump php-debugbar by @barryvdh in https://github.com/barryvdh/laravel-debugbar/pull/1791
+* Fix counter tests by @barryvdh in https://github.com/barryvdh/laravel-debugbar/pull/1792
+* `$group` arg support on TimelineCollectors methods by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1789
+* Collect other eloquent model events by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1781
+* Add new cache events on CacheCollector by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1773
+* Exclude events on EventCollector by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1786
+* Use `addWarning` on warnings, silenced errors, notices by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1767
+* Do not rely on DB::connection() to get information in query collector by @cweiske in https://github.com/barryvdh/laravel-debugbar/pull/1779
+* Trace file for Gate checks(GateCollector) by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1770
+* Fix support for PDOExceptions by @LukeTowers in https://github.com/barryvdh/laravel-debugbar/pull/1752
+* Time measure on cache events by @erikn69 in https://github.com/barryvdh/laravel-debugbar/pull/1794
+* fix debugbar for Lumen usage by @flibidi67 in https://github.com/barryvdh/laravel-debugbar/pull/1796
+* Custom path for Inertia views by @joaopms in https://github.com/barryvdh/laravel-debugbar/pull/1797
+* Better contrast in dark theme titles. by @angeljqv in https://github.com/barryvdh/laravel-debugbar/pull/1798
+
+### New Contributors
+
+* @wimski made their first contribution in https://github.com/barryvdh/laravel-debugbar/pull/1784
+* @cweiske made their first contribution in https://github.com/barryvdh/laravel-debugbar/pull/1779
+* @flibidi67 made their first contribution in https://github.com/barryvdh/laravel-debugbar/pull/1796
+* @joaopms made their first contribution in https://github.com/barryvdh/laravel-debugbar/pull/1797
+
+**Full Changelog**: https://github.com/barryvdh/laravel-debugbar/compare/v3.15.4...v3.16.0
+
 ## v3.15.4 - 2025-04-16
 
 ### What's Changed